requirements.txt
torch==2.1.2
torchvision==0.16.2
onnx>=1.12.0
onnx-tf==1.10.0
tensorflow==2.13.0
tensorflow-probability==0.21.0
protobuf<4
Pillow==11.3.0
matplotlib==3.9.4